requirements.txt
django==3.2.16
environs==9.5.0
python-telegram-bot==13.*
